% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_rmd_template.R
\name{copy_rmd_template}
\alias{copy_rmd_template}
\title{Create a new directory with a clean RMarkdown file}
\usage{
copy_rmd_template(fname = "new-doc")
}
\arguments{
\item{fname}{name of the Rmd file to create}
}
\description{
This function creates a new subdirectory inside the current directory, which will
contain a ready-to-use RMarkdown file to be rendered in the given format.
}
