% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_process.R
\name{extract_gff3_attr}
\alias{extract_gff3_attr}
\title{extract gff3 attribute}
\usage{
extract_gff3_attr(gff_attributes)
}
\arguments{
\item{gff_attributes}{attributes. character vector}
}
\description{
extract gff3 attribute
}
