% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gff.R
\name{filter_feature}
\alias{filter_feature}
\title{Filtering specified feature from gff}
\usage{
filter_feature(gff, feature = "exon")
}
\arguments{
\item{gff}{gff data frame}

\item{feature}{feature}
}
\description{
Filtering specified feature from gff
}
