% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gff.R
\name{filter_repr_model}
\alias{filter_repr_model}
\title{Filtering row by representative model from given data.frame}
\usage{
filter_repr_model(df, rep_model, rep_model_col = NA_character_)
}
\arguments{
\item{df}{data frame}

\item{rep_model}{character vector}

\item{rep_model_col}{column name}
}
\description{
Filtering row by representative model from given data.frame
}
