% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_process.R
\name{multiply_overlap_feature}
\alias{multiply_overlap_feature}
\title{Multiply overlaped feature rows}
\usage{
multiply_overlap_feature(gff)
}
\arguments{
\item{gff}{gff3 data.frame}
}
\description{
Multiply overlaped feature rows
}
