% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mydata_urls.R
\docType{data}
\name{mydata_urls}
\alias{mydata_urls}
\title{mydata_urls}
\format{A named list with 6 variables
\describe{
  \item{tair_gff3_path}{}
  \item{tair_gff3_aip_path}{}
  \item{araport_gtf_path}{}
  \item{araport_gff3_path}{}
  \item{tair_func_desc_path}{}
  \item{tair_repr_model_path}{}
}}
\usage{
mydata_urls
}
\description{
mydata_urls
}
\keyword{datasets}
