% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_process.R
\name{split_attr_col}
\alias{split_attr_col}
\title{Splitting gff3 attribute column}
\usage{
split_attr_col(gff_attributes)
}
\arguments{
\item{gff_attributes}{attributes. character vector}
}
\description{
Splitting gff3 attribute column
}
